Este pacote atua como uma **CLI** e **engine** para:***Escalonamento de Agentes**: Agende execuções programadas de agentes (via Flask/Celery) diretamente do terminal.* **Scaffold de Projetos**: Crie esqueletos completos de aplicações Python (Flask) com templates e containers Docker
