+++ SoftwareAISkeletonEngine/Package.json
@@ -1,17 +1,21 @@
 {
-    "name": "softwareai-skeleton-engine",
+    "name": "@ualers/softwareai-skeleton-engine",
     "type": "module",
     "version": "1.0.0",
     "description": "CLI to scaffold Python application projects",
+    "publishConfig": {
+        "access": "public"
+    },
     "bin": {
-      "create-py-app": "./index.js"
+        "create-py-app": "index.js"
     },
     "dependencies": {
-      "commander": "^11.0.0",
-      "fs-extra": "^11.1.1",
-      "inquirer": "^9.2.7"
+        "commander": "^11.0.0",
+        "fs-extra": "^11.1.1",
+        "inquirer": "^9.2.7",
+        "axios": "^1.4.0",
+        "moment-timezone": "^0.5.43"
     },
-    "author": "",
+    "author": "ualers",
     "license": "MIT"
-  }
-  
+}
